fix(starterkit): give each converter asset the file Unity needs
- Unity only builds a MonoScript for the type whose name matches the file, so
  the assets declared together in ConverterAssets.cs could never be loaded.
- Annotate the converter type arguments as nullable, dropping the null-forgiving
  casts an empty asset needed to return the default.
- Report a missing asset on every conversion instead of only the first.
