+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Converters/Assets/BoolConverterAsset.cs
@@ -1,0 +1,12 @@
+#nullable enable
+using UnityEngine;
+
+// ReSharper disable once CheckNamespace
+namespace Aspid.MVVM.StarterKit
+{
+    /// <summary>
+    /// <see cref="ConverterAsset{TFrom, TTo}"/> for <see cref="bool"/> values.
+    /// </summary>
+    [CreateAssetMenu(menuName = "Aspid/MVVM/Converters/Bool Converter", fileName = "BoolConverter")]
+    public sealed class BoolConverterAsset : ConverterAsset<bool, bool> { }
+}
