+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Converters/Assets/BoolConverterAsset.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: f6516a776cd74a74ae58d97d3222782b
+timeCreated: 1786899180
