+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Converters/Assets/ColorConverterAsset.cs
@@ -1,0 +1,12 @@
+#nullable enable
+using UnityEngine;
+
+// ReSharper disable once CheckNamespace
+namespace Aspid.MVVM.StarterKit
+{
+    /// <summary>
+    /// <see cref="ConverterAsset{TFrom, TTo}"/> for <see cref="Color"/> values.
+    /// </summary>
+    [CreateAssetMenu(menuName = "Aspid/MVVM/Converters/Color Converter", fileName = "ColorConverter")]
+    public sealed class ColorConverterAsset : ConverterAsset<Color, Color> { }
+}
