+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Converters/Assets/ColorConverterAsset.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 1e8f52b3dde94bd49e55eef8a91ad10f
+timeCreated: 1786899190
