+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Converters/Assets/FloatConverterAsset.cs
@@ -1,0 +1,12 @@
+#nullable enable
+using UnityEngine;
+
+// ReSharper disable once CheckNamespace
+namespace Aspid.MVVM.StarterKit
+{
+    /// <summary>
+    /// <see cref="ConverterAsset{TFrom, TTo}"/> for <see cref="float"/> values.
+    /// </summary>
+    [CreateAssetMenu(menuName = "Aspid/MVVM/Converters/Float Converter", fileName = "FloatConverter")]
+    public sealed class FloatConverterAsset : ConverterAsset<float, float> { }
+}
