+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Converters/Assets/FloatConverterAsset.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: d2755e4c443049769581c950486e8b79
+timeCreated: 1786899138
