+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Converters/Assets/ObjectToStringConverterAsset.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 4edee24fe7bb449aa39012793234dd91
+timeCreated: 1786898841
