+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Converters/Assets/StringConverterAsset.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: c725f7c06eb8444f992abcda7cb7d5c2
+timeCreated: 1786899092
